pyproject.toml
[project]
name = "hterm"
version = "0.1.0"
description = "A custom terminal"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paramiko>=3.5.1",
    "pyfiglet>=1.0.2",
    "pyserial>=3.5",
    "pyside6>=6.8.2.1",
    "pywinpty>=2.0.15; sys_platform == 'win32'",
    "pyyaml>=6.0.2",
    "qtawesome>=1.3.1",
]

[dependency-groups]
dev = [
    "pillow>=11.1.0",
    "pyinstaller>=6.12.0",
]
